% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_helper_functions.R
\name{carryForwardNA}
\alias{carryForwardNA}
\title{carryForwardNA}
\usage{
carryForwardNA(thelist)
}
\arguments{
\item{thelist}{the list to check for NA head}
}
\description{
Checks if front of list is NA, then replaces it with the most recent value which is not NA.
}
