% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_helper_functions.R
\name{dbinom_APA}
\alias{dbinom_APA}
\title{Arbitrary Precision Binomial density function.}
\usage{
dbinom_APA(x, size, prob, precBits = 128)
}
\arguments{
\item{x}{quantile (number of Bernoulli successes with success probability prob)}

\item{size}{number of Bernoulli trials to perform with success probability prob}

\item{prob}{probability of success for each independent Bernoulli trial}

\item{precBits}{number of bits of precision}
}
\description{
Binomial density function with arbitrary precision. Uses library Rmpfr for arbitrary precision arithmetic.
}
\examples{
# compare apa to non apa dpois:
sum(dbinom_APA(x = 0:10, size = 10, prob=0.2, precBits = 256))
sum(dbinom(x = 0:10, size = 10, prob=0.2))
}
