% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_helper_functions.R
\name{dpois_APA}
\alias{dpois_APA}
\title{Arbitrary Precision Poisson density function.}
\usage{
dpois_APA(x, lambda, precBits = 128)
}
\arguments{
\item{x}{quantile}

\item{lambda}{poisson mean parameter}

\item{precBits}{number of bits of precision}
}
\description{
Poisson density function with arbitrary precision. Uses library Rmpfr for arbitrary precision arithmetic.
}
\examples{
# compare apa to non apa dpois:
sum(dpois_APA(x = 0:70, lambda = 10, precBits = 128))
sum(dpois(x = 0:40, lambda = 10)) # sums to 1, indicating that dpois(x=41, lambda=10) should be zero
dpois(x=41, lambda=10) # is not zero
dpois_APA(x=41, lambda=10, precBits = 128)
}
