% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineSearch_NAPA.R
\name{lineSearch_NAPA}
\alias{lineSearch_NAPA}
\title{lineSearch_NAPA}
\usage{
lineSearch_NAPA(x_curr, dk, func, grad_Fx = NULL,
  lineSearchMaxSteps = 100, ...)
}
\arguments{
\item{x_curr}{starting x value}

\item{dk}{direction vector to search along}

\item{func}{function to perform line search on}

\item{grad_Fx}{(defaults to NULL) if not NULL, the gradient of func at x_curr}

\item{lineSearchMaxSteps}{(defaults to 100) maximum number of iterations before stopping}

\item{...}{extra parameters passed on to func}
}
\description{
line searching algorithm used inside non-APA optimization algorithms.
}
\examples{
# simple 1D quadratic function optimization
lineSearch_NAPA(x_curr = 1, dk = -0.2, lineSearchMaxSteps = 100, func = function(x) {x^2})

# simple 2D quadratic function optimization
func2 <- function(par, centerx=0, centery=0) {
  (par[1]-centerx)^2+(par[2]-centery)^2
}
lineSearch_NAPA(x_curr = c(.015,-.015), dk = c(-0.2,0.2), func=func2)
}
