% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_helper_functions.R
\name{plogis_APA}
\alias{plogis_APA}
\title{Arbitrary Precision Logistic density function.}
\usage{
plogis_APA(x, location = 0, scale = 1, precBits = 128)
}
\arguments{
\item{x}{quantile}

\item{location}{Location parameter (default 0)}

\item{scale}{Scale parameter (default 1)}

\item{precBits}{Number of bits of precision}
}
\description{
Logistic density function with arbitrary precision. Uses library Rmpfr for arbitrary precision arithmetic.
}
