% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_helper_functions.R
\name{plotConvergence}
\alias{plotConvergence}
\title{plotConvergence}
\usage{
plotConvergence(
  optimOutput,
  variable = NULL,
  digits = 5,
  labels = T,
  flip_axes = F
)
}
\arguments{
\item{optimOutput}{The result of running an optimization algorithm with keepValues=TRUE}

\item{variable}{Default is NULL, will print convergence plot for all variables, otherwise an integer vector (such as c(1,3)) indicating which variables to print convergence plot for.}

\item{digits}{Default is 5. Number of digits to print on convergence plot.}

\item{labels}{If TRUE, print function values as labels on the plot.}

\item{flip_axes}{If TRUE, will plot the convergence horizontally rather than vertically.}
}
\description{
Plot the path to convergence of an optimization algorithm (such as optim_DFP_APA).
}
\examples{
# N dimensional quadratic
funcND <- function(par, center, precBits=64) {
  par <- Rmpfr::mpfr(par, precBits)
  log(sum(exp((par-center)^2)))
} 
opt <- optimizeAPA::optim_DFP_APA(starts = c(0,0,0), func = funcND, center=c(1,2,3), keepValues=TRUE)
plotConvergence(opt, variable=c(1,3), labels=F)
}
