% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_helper_functions.R
\name{updateList}
\alias{updateList}
\title{updateList}
\usage{
updateList(new_el, old_list = list(), M = 1)
}
\arguments{
\item{new_el}{element to add to top of list}

\item{old_list}{list to append new_el to}

\item{M}{maximum size of list (will truncate by removing bottom if length is greater than M)}
}
\description{
Adds new element to top of list, removing bottom elements if new length is greater tham M
}
