% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_helper_functions.R
\name{plotConvergence}
\alias{plotConvergence}
\title{plotConvergence}
\usage{
plotConvergence(optimOutput, variable = NULL, digits = 5)
}
\arguments{
\item{optimOutput}{The result of running an optimization algorithm with keepValues=TRUE}

\item{variable}{Default is NULL, will print convergence plot for all variables, otherwise an integer vector (such as c(1,3)) indicating which variables to print convergence plot for.}

\item{digits}{Default is 5. Number of digits to print on convergence plot.}
}
\description{
Plot the path to convergence of an optimization algorithm (such as optim_DFP_APA).
}
\examples{
# N dimensional quadratic
funcND <- function(par, center, precBits=64) {
  par <- Rmpfr::mpfr(par, precBits)
  sum((par-center)^2)
} 
opt <- optimizeAPA::optim_DFP_APA(starts = c(0,0,0), func = funcND, center=c(1,2,3), keepValues=TRUE)
plotConvergence(opt, variable=c(1,3))
}
